requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
authlib==1.4.0
httpx==0.28.1
itsdangerous==2.2.0
jinja2==3.1.4
python-multipart==0.0.19
aioimaplib==2.0.1
aiosmtplib==3.0.2
bleach==6.2.0
starlette-csrf==3.0.0
pydantic-settings==2.7.0
